% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_duplicates_mark.R
\name{add_equality_group_number}
\alias{add_equality_group_number}
\title{add_equality_group_number}
\usage{
add_equality_group_number(x, partner_list)
}
\arguments{
\item{x}{c14_date_list}

\item{partner_list}{partner list produced by generate_list_of_equality_partners()}
}
\value{
c14_date_list with additional column duplicate_group
}
\description{
add_equality_group_number
}
\keyword{internal}
