% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_duplicates_mark.R
\name{add_equality_group_number_based_on_equality_partner_list}
\alias{add_equality_group_number_based_on_equality_partner_list}
\title{add column with equality group number to c14_date_list}
\usage{
add_equality_group_number_based_on_equality_partner_list(x, partner_list)
}
\arguments{
\item{x}{c14_date_list}

\item{partner_list}{partner list produced by generate_list_of_equality_partners()}
}
\value{
c14_date_list with additional column duplicate_group
}
\description{
add column with equality group number to c14_date_list
}
