% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{add_or_replace_column_in_df}
\alias{add_or_replace_column_in_df}
\title{add_or_replace_column_in_df}
\usage{
add_or_replace_column_in_df(x, column_name_s, column_content_mi, ...)
}
\arguments{
\item{x}{data.frame}

\item{column_name_s}{name of new column}

\item{column_content_mi}{content of new column}

\item{...}{passed to tibble::add_column()}
}
\value{
data.frame with new column
}
\description{
add_or_replace_column_in_df
}
\keyword{internal}
