% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_basic.R
\name{as.c14_date_list}
\alias{as.c14_date_list}
\title{Convert an object of class data.frame or tibble to class c14_date_list}
\usage{
as.c14_date_list(x, ...)
}
\arguments{
\item{x}{a variable}

\item{...}{further arguments passed to or from other methods}
}
\value{
an object of class c14_date_list
}
\description{
If an object is of class data.frame or tibble (tbl & tbl_df), it can be
converted to an object of class c14_date_list
}
