% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_convert.R
\name{as.sf}
\alias{as.sf}
\alias{as.sf.default}
\alias{as.sf.c14_date_list}
\title{convert a c14_date_list to a sf object}
\usage{
as.sf(x)

\method{as.sf}{default}(x)

\method{as.sf}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class sf
}
\description{
convert a c14_date_list to a sf object
}
