% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14databases.R
\docType{package}
\name{c14databases-package}
\alias{c14databases}
\alias{c14databases-package}
\title{c14databases: Download and Prepare C14 Dates from Different Source Databases}
\description{
Query different C14 date databases and apply basic data cleaning, merging and calibration steps.
}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} [copyright holder]

}
\keyword{internal}
