% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_calibrate.R
\name{calibrate}
\alias{calibrate}
\alias{calibrate.default}
\alias{calibrate.c14_date_list}
\title{Calibrate dates}
\usage{
calibrate(x, ...)

\method{calibrate}{default}(x, ...)

\method{calibrate}{c14_date_list}(x, choices = c("sigmarange"), sigma = 2,
  ...)
}
\arguments{
\item{x}{an object of class c14_date_list}

\item{...}{Arguments to be passed to methods for c14_date_list}

\item{choices}{whether the result should include the full calibrated probability dataframe ('probdist')
or the sigma range ('sigmarange'). Both arguments may be given at the same time.}

\item{sigma}{the desired sigma value (1,2,3) for the calibrated sigma ranges}
}
\value{
an object of class c14_date_list
}
\description{
Calibrate all valid dates in a c14_date_list
}
