% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_calibrate.R
\name{calibrate_to_probability_distribution}
\alias{calibrate_to_probability_distribution}
\title{calibrate_to_probability_distribution}
\usage{
calibrate_to_probability_distribution(x, ...)
}
\arguments{
\item{x}{c14_date_list}

\item{...}{passed to Bchron::BchronCalibrate()}
}
\value{
list with probability distribution data frames
}
\description{
calibrate_to_probability_distribution
}
