% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{check_connection_to_url}
\alias{check_connection_to_url}
\title{check_connection_to_url}
\usage{
check_connection_to_url(db_url)
}
\arguments{
\item{db_url}{url string}
}
\value{
logical
}
\description{
check_connection_to_url
}
\keyword{internal}
