% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{check_if_columns_are_present}
\alias{check_if_columns_are_present}
\title{check_if_columns_are_present}
\usage{
check_if_columns_are_present(x, columns)
}
\arguments{
\item{x}{c14_date_list}

\item{columns}{name of columns column}
}
\value{
NULL - called for side effect stop()
}
\description{
check_if_columns_are_present
}
