% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_helper_functions.R
\name{check_if_packages_are_available}
\alias{check_if_packages_are_available}
\title{check_if_packages_are_available}
\usage{
check_if_packages_are_available(packages_ch)
}
\arguments{
\item{packages_ch}{packages that should be available}
}
\value{
NULL - called for side effect stop()
}
\description{
check_if_packages_are_available
}
