% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_precision.R
\name{circumference_calculator}
\alias{circumference_calculator}
\title{circumference_calculator}
\usage{
circumference_calculator(x, mode)
}
\arguments{
\item{x}{vector of latitude or longitude coordinates}

\item{mode}{a character "lat" or "lon"}
}
\value{
vecor with circumference values at specific latitudes
}
\description{
circumference_calculator
}
