% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_classify_material.R
\name{classify_material}
\alias{classify_material}
\alias{classify_material.default}
\alias{classify_material.c14_date_list}
\title{Apply material classification}
\usage{
classify_material(x, material_thesaurus = c14bazAAR::get_material_thesaurus(),
  quiet = FALSE)

\method{classify_material}{default}(x,
  material_thesaurus = c14bazAAR::get_material_thesaurus(), quiet = FALSE)

\method{classify_material}{c14_date_list}(x,
  material_thesaurus = c14bazAAR::get_material_thesaurus(), quiet = FALSE)
}
\arguments{
\item{x}{an object of class c14_date_list}

\item{material_thesaurus}{a thesaurus table (default: c14bazAAR::material_thesaurus)}

\item{quiet}{suppress printed output}
}
\value{
an object of class c14_date_list
}
\description{
Add column material_thes with simplified and unified terms
}
