% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list.R
\name{clean}
\alias{clean}
\alias{clean.default}
\alias{clean.c14_date_list}
\title{Clean dataset}
\usage{
clean(x)

\method{clean}{default}(x)

\method{clean}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Apply some data cleaning steps to a c14_date_list
}
