% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_general.R
\name{clean_latlon}
\alias{clean_latlon}
\title{clean dataset}
\usage{
clean_latlon(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Apply some data cleaning steps to a c14_date_list
}
\keyword{internal}
