% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_precision.R
\name{coordinate_precision}
\alias{coordinate_precision}
\alias{coordinate_precision.default}
\alias{coordinate_precision.c14_date_list}
\title{Return coordinate precision according to number length}
\usage{
coordinate_precision(x)

\method{coordinate_precision}{default}(x)

\method{coordinate_precision}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Return coordinate precision according to number length
}
