% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_country_attribution.R
\name{country_attribution}
\alias{country_attribution}
\alias{determine_country_by_coordinate}
\alias{determine_country_by_coordinate.default}
\alias{determine_country_by_coordinate.c14_date_list}
\title{Determine the country of all dates in a \strong{c14_date_list} from their coordinates}
\usage{
determine_country_by_coordinate(x, suppress_spatial_warnings = TRUE)

\method{determine_country_by_coordinate}{default}(x, suppress_spatial_warnings = TRUE)

\method{determine_country_by_coordinate}{c14_date_list}(x, suppress_spatial_warnings = TRUE)
}
\arguments{
\item{x}{an object of class c14_date_list}

\item{suppress_spatial_warnings}{suppress some spatial data messages and warnings}
}
\value{
an object of class c14_date_list with the additional column \strong{country_coord}
}
\description{
\code{c14bazAAR::determine_country_by_coordinate()} adds the column
\strong{country_coord} with standardized country attribution based on the coordinate
information for the dates.
Due to the inconsistencies in the \strong{country} column in many c14 source databases
it's often necessary to rely on the coordinate position (\strong{lat} & \strong{lon})
for country attribution information. Unfortunately not all source databases store
coordinates.
}
\examples{
library(magrittr)
example_c14_date_list \%>\%
  determine_country_by_coordinate()

}
