% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{db_info_table}
\alias{db_info_table}
\title{Database lookup table}
\format{
a data.frame. Columns:
\itemize{
 \item{db: database name}
 \item{version: database version}
 \item{url_num: url number (some databases are spread over multiple files)}
 \item{url: file url where the database can be downloaded}
}
}
\description{
Lookup table for general source database information.
}
\seealso{
Other lookup_tables: 
\code{\link{country_thesaurus}},
\code{\link{material_thesaurus}}
}
\concept{lookup_tables}
