% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_calibrate.R
\name{determine_calage_from_probability_distribution}
\alias{determine_calage_from_probability_distribution}
\title{determine_calage_from_probability_distribution}
\usage{
determine_calage_from_probability_distribution(x)
}
\arguments{
\item{x}{c14_date_list}
}
\value{
vector of calages
}
\description{
determine_calage_from_probability_distribution
}
