% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/c14_date_list_spatial_determine_country_by_coordinate.R
\name{determine_country_by_coordinate}
\alias{determine_country_by_coordinate}
\alias{determine_country_by_coordinate.default}
\alias{determine_country_by_coordinate.c14_date_list}
\title{Determine country by coordinate in a \strong{c14_date_list}}
\usage{
determine_country_by_coordinate(x)

\method{determine_country_by_coordinate}{default}(x)

\method{determine_country_by_coordinate}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list with the additional column
\strong{country_coord}.
}
\description{
Add column \strong{country_coord} with standardized country
attribution based on the coordinate information of the dates. \cr
Due to the inconsistencies in the \strong{country} column
in many c14 source databases (see \code{c14bazAAR::standardize_country_name()}),
it's often necessary to rely on the coordinate position (\strong{lat} &
\strong{lon}) for reliable country attribution information.
}
