% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_calibrate.R
\name{determine_dates_out_of_range_of_calcurve}
\alias{determine_dates_out_of_range_of_calcurve}
\title{determine_dates_out_of_range_of_calcurve}
\usage{
determine_dates_out_of_range_of_calcurve(x)
}
\arguments{
\item{x}{c14_date_list}
}
\value{
indizes of dates in c14_date_list which are out of calcurve range
}
\description{
determine_dates_out_of_range_of_calcurve
}
\keyword{internal}
