% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_precision.R
\name{digits_counter}
\alias{digits_counter}
\title{digits_counter}
\usage{
digits_counter(x, ...)
}
\arguments{
\item{x}{vector}
}
\value{
vecor with numer of digits
}
\description{
digits_counter
}
