% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_duplicates_mark.R
\name{duplicates_mark}
\alias{duplicates_mark}
\alias{duplicates_mark.default}
\alias{duplicates_mark.c14_date_list}
\title{mark duplicates}
\usage{
duplicates_mark(x)

\method{duplicates_mark}{default}(x)

\method{duplicates_mark}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list with the additional column duplicate
}
\description{
mark double entries in a c14_date_list by
comparing the Labcodes
}
