% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_enforce_types.R
\name{enforce_types}
\alias{enforce_types}
\alias{enforce_types.default}
\alias{enforce_types.c14_date_list}
\title{Enforce variable types}
\usage{
enforce_types(x)

\method{enforce_types}{default}(x)

\method{enforce_types}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Enforce variable types in a c14_date_list and remove everything that doesn't fit
}
