% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_estimate_statial_quality.R
\name{estimate_spatial_quality}
\alias{estimate_spatial_quality}
\alias{estimate_spatial_quality.default}
\alias{estimate_spatial_quality.c14_date_list}
\title{Estimate quality of coordinate information}
\usage{
estimate_spatial_quality(x,
  country_thesaurus_table = c14databases::country_thesaurus)

\method{estimate_spatial_quality}{default}(x,
  country_thesaurus_table = c14databases::country_thesaurus)

\method{estimate_spatial_quality}{c14_date_list}(x,
  country_thesaurus_table = c14databases::country_thesaurus)
}
\arguments{
\item{x}{an object of class c14_date_list}

\item{country_thesaurus_table}{a thesaurus table (default: c14databases::country_thesaurus)}
}
\value{
an object of class c14_date_list
}
\description{
Estimate quality of coordinates with a set of tests
}
