% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_c14_date_list}
\alias{example_c14_date_list}
\title{Example c14_date_list}
\format{A c14_date_list. See variable_reference for an explanation of
the variable meaning.}
\description{
c14_date_list with 200 random dates for tests and example code.
}
\concept{c14_date_lists}
