% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_finalize_country_name.R
\name{finalize_country_name}
\alias{finalize_country_name}
\alias{finalize_country_name.default}
\alias{finalize_country_name.c14_date_list}
\title{Chooses the country_name, that is most likely to be correct}
\usage{
finalize_country_name(x)

\method{finalize_country_name}{default}(x)

\method{finalize_country_name}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Picks the country name in a hierarchical order from the results
of the determine_country_by_coords and standardize_country_name functions,
followed by the original input of the database
}
