% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_thesauri.R
\name{find_lookup_decisions}
\alias{find_lookup_decisions}
\title{find_lookup_decisions}
\usage{
find_lookup_decisions(x, variants_column, corrected_column, thesaurus)
}
\arguments{
\item{x}{a c14_date_list}

\item{variants_column}{name of the column with heterogeneous values}

\item{corrected_column}{name of the column with the corrected values}

\item{thesaurus}{data.frame with correct values (cor) and variants (var)}
}
\value{
a tibble with the country names and the new country_thes names
found by \code{find_correct_name_by_stringdist_comparison()}
}
\description{
find_lookup_decisions
}
\keyword{internal}
