% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_basic.R
\name{format.c14_date_list}
\alias{format.c14_date_list}
\title{Encode a c14_date_list in a Common Format}
\usage{
\method{format}{c14_date_list}(x, ...)
}
\arguments{
\item{x}{a c14_date_list}

\item{...}{further arguments passed to or from other methods}
}
\value{
A string representation of the c14_date_list
}
\description{
Format an c14_date_list for pretty printing
}
