% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_fuse.R
\name{fuse}
\alias{fuse}
\alias{fuse.default}
\alias{fuse.c14_date_list}
\title{fuse date tables}
\usage{
fuse(...)

\method{fuse}{default}(...)

\method{fuse}{c14_date_list}(...)
}
\arguments{
\item{...}{objects of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
rowbind multiple c14_date_lists
}
