% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_duplicates_mark.R
\name{generate_list_of_equality_partners}
\alias{generate_list_of_equality_partners}
\title{search for equality partners in vector}
\usage{
generate_list_of_equality_partners(x)
}
\arguments{
\item{x}{vector}
}
\value{
list of unique partners
}
\description{
search for equality partners in vector
}
