% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_dates.R
\name{get_all_dates}
\alias{get_all_dates}
\title{get all available data}
\usage{
get_all_dates()
}
\description{
Downloads the current version of several databases and merges them to one c14_date_list
}
\examples{

\dontrun{
  all_dates <- get_all_dates()
}

}
