% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_dates.R
\name{get_all_parser_functions}
\alias{get_all_parser_functions}
\title{get_all_parser_functions}
\usage{
get_all_parser_functions()
}
\value{
vector with all parser functions in c14bazAAR
}
\description{
get_all_parser_functions
}
