% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austarch.R
\name{get_AustArch}
\alias{get_AustArch}
\title{get current AustArch-Database}
\usage{
get_AustArch()
}
\description{
Downloads the current version of the AustArch-Database from \url{http://archaeologydataservice.ac.uk/archives/view/austarch_na_2014/}.
}
\examples{

\dontrun{
  CARD <- get_CARD()
}

}
