% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calpal.R
\name{get_CALPAL}
\alias{get_CALPAL}
\title{get current CALPAL-Database}
\usage{
get_CALPAL()
}
\description{
Downloads the current version of the CALPAL-Database from \url{https://github.com/nevrome/CalPal-Database}.
}
\examples{

\dontrun{
  CALPAL <- get_CALPAL()
}

}
