% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/card.R
\name{get_CARD}
\alias{get_CARD}
\title{get current CARD-Database [no open API]}
\usage{
get_CARD()
}
\description{
Downloads the current version of the CARD-Database from \url{http://www.canadianarchaeology.ca/}.
}
\examples{

\dontrun{
  CARD <- get_CARD()
}

}
