% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_dbs.R
\name{get_db_url}
\alias{get_db_url}
\title{get db url}
\usage{
get_db_url(db_name,
  ref_url = paste(c("https://raw.githubusercontent.com", "ropensci",
  "c14bazAAR", "master", "data-raw", "url_reference.csv"), collapse = "/"))
}
\arguments{
\item{db_name}{name of the database}

\item{ref_url}{url of the relevant reference table}
}
\description{
Downloads information for c14 source databases from a reference table
on github.
}
