% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euroevol.R
\name{get_EUROEVOL}
\alias{get_EUROEVOL}
\title{get current EUROEVOL-Database}
\usage{

  get_EUROEVOL(db_url1 = "http://discovery.ucl.ac.uk/1469811/7/EUROEVOL09-07-201516-34_C14Samples.csv",
  db_url2 = "http://discovery.ucl.ac.uk/1469811/9/EUROEVOL09-07-201516-34_CommonSites.csv",
  db_url3 = "http://discovery.ucl.ac.uk/1469811/8/EUROEVOL09-07-201516-34_CommonPhases.csv")
}
\arguments{
\item{db_url1}{string with weblink to c14 archive file}

\item{db_url2}{string with weblink to c14 archive file}

\item{db_url3}{string with weblink to c14 archive file}
}
\description{
Downloads the current version of the EUROEVOL-Database from \url{http://discovery.ucl.ac.uk/1469811/}.
}
\examples{

\dontrun{
  EUROEVOL <- get_EUROEVOL()
}

}
