% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eastafrica.R
\name{get_KITEeastAfrica}
\alias{get_KITEeastAfrica}
\title{get KITE East Africa v2.1.csv from ...}
\usage{
get_KITEeastAfrica(db_url = get_db_url("KITEeastAfrica"))
}
\arguments{
\item{db_url}{string with weblink to c14 archive file}
}
\description{
See:
- A Database of Radiocarbon Dates for Palaeoenvironmental Research in Eastern Africa: https://www.openquaternary.com/articles/10.5334/oq.22/
- Radiocarbon dates from eastern Africa in the CARD2.0 format: https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/NJLNRJ
}
\details{
The dataset is stored within https://dataverse.harvard.edu that requires the dataverse-package (see https://cran.r-project.org/web/packages/dataverse/vignettes/C-retrieval.html)
}
\examples{

\dontrun{
  KITEeastAfrica <- get_KITEeastAfrica()
}

}
