% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radon.R
\name{get_RADON}
\alias{get_RADON}
\title{get current RADON-Database}
\usage{
get_RADON()
}
\description{
Downloads the current version of the RADON-Database from \url{http://radon.ufg.uni-kiel.de/}.
}
\examples{

\dontrun{
  RADON <- get_RADON()
}

}
