% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radonb.R
\name{get_RADONB}
\alias{get_RADONB}
\title{get current RADONB-Database}
\usage{
get_RADONB(db_url = get_db_url("RADON-B"))
}
\arguments{
\item{db_url}{string with weblink to c14 archive file}
}
\description{
Downloads the current version of the RADON-Database from \url{http://radon-b.ufg.uni-kiel.de/}.
}
\examples{

\dontrun{
  RADONB <- get_RADONB()
}

}
