% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_thesauri.R
\name{get_thesaurus}
\alias{get_thesaurus}
\title{get_thesaurus}
\usage{
get_thesaurus(url)
}
\arguments{
\item{url}{}
}
\value{
thesaurus data.frame
}
\description{
helper function to download thesaurus files
}
