% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_calibrate.R
\name{hdr}
\alias{hdr}
\title{Calculate highest density regions for Bchron calibrated ages}
\usage{
hdr(calprobdistr, prob = 0.95)
}
\arguments{
\item{calprobdistr}{the probability distribution}

\item{prob}{The desired probability interval, in the range(0, 1)}
}
\value{
a dataframe containing the hdr
}
\description{
A function for computing highest density regions (HDRs).
Adopted from the Bchron package.
}
\details{
The output of this function is a list of contiguous ranges which
cover the probability interval requested. A highest density region might
have multiple such ranges if the calibrated date is multi-modal. These
differ from credible intervals, which are always contiguous but will not
be a good representation of a multi-modal probability distribution.
}
\keyword{internal}
