% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_precision.R
\name{individual_precision}
\alias{individual_precision}
\title{latitude_precision}
\usage{
individual_precision(x, ...)
}
\arguments{
\item{x}{vector of latitude}
}
\value{
vecor with precision in meters
}
\description{
latitude_precision
}
