% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_thesauri.R
\name{inspect_lookup_country}
\alias{inspect_lookup_country}
\alias{inspect_lookup_material}
\title{Get an overview of the lookup decisions}
\usage{
inspect_lookup_country(x, thesaurus = c14bazAAR::country_thesaurus)

inspect_lookup_material(x, thesaurus = c14bazAAR::material_thesaurus)
}
\arguments{
\item{x}{a c14_date_list with the columns country and country_thes or material
and material_thes which went through either or both \link{fix_database_country_name}
or \link{classify_material}}

\item{thesaurus}{see \link{fix_database_country_name} or \link{classify_material}}
}
\value{
a data.frame with information about lookup decisions
}
\description{
These functions allow to inspect the lookups done in \link{fix_database_country_name}
and \link{classify_material}. This can be useful to understand which values are not
covered in the in the thesaurus tables.
}
