% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_basic.R
\name{is.c14_date_list}
\alias{is.c14_date_list}
\title{Check if an object is of class c14_date_list}
\usage{
is.c14_date_list(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{further arguments passed to or from other methods}
}
\value{
true if x is a c14_date_list, false otherwise
}
\description{
Check if an object is of class c14_date_list
}
