% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_classify_material.R
\name{lookup_in_thesaurus_table}
\alias{lookup_in_thesaurus_table}
\title{lookup_in_thesaurus_table}
\usage{
lookup_in_thesaurus_table(x, thesaurus_df)
}
\arguments{
\item{x}{vector of values to look up in thesaurus}

\item{thesaurus_df}{reference table that contains variants and correct values}
}
\value{
a vector with the correct values
}
\description{
lookup_in_thesaurus_table
}
\keyword{internal}
