% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_duplicates_mark.R
\name{mark_duplicates}
\alias{mark_duplicates}
\alias{mark_duplicates.default}
\alias{mark_duplicates.c14_date_list}
\title{Mark duplicates in a \strong{c14_date_list}}
\usage{
mark_duplicates(x)

\method{mark_duplicates}{default}(x)

\method{mark_duplicates}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list with the additional
column \strong{duplicate_group}
}
\description{
Duplicates are found by comparision of \strong{labnr}s.
Only dates with exactly equal \strong{labnr}s are considered duplicates.
Duplicate groups are numbered (from 0) and these numbers linked to
the individual dates in the new column \strong{duplicate_group}.
Duplicates can be removed with \code{c14bazAAR::remove_duplicates()}.
}
\examples{
mark_duplicates(example_c14_date_list)

}
