% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_material_classification.R
\name{material_classification}
\alias{material_classification}
\alias{material_classification.default}
\alias{material_classification.c14_date_list}
\title{Apply material classification}
\usage{
material_classification(x,
  material_thesaurus_df = c14bazAAR::material_thesaurus)

\method{material_classification}{default}(x,
  material_thesaurus_df = c14bazAAR::material_thesaurus)

\method{material_classification}{c14_date_list}(x,
  material_thesaurus_df = c14bazAAR::material_thesaurus)
}
\arguments{
\item{x}{an object of class c14_date_list}

\item{material_thesaurus_df}{a thesaurus table (default: c14bazAAR::material_thesaurus)}
}
\value{
an object of class c14_date_list
}
\description{
Add column material_thes with simplified and unified terms
}
