% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{material_thesaurus}
\alias{material_thesaurus}
\title{Material thesaurus}
\format{
...
}
\description{
Lookup table for material categories.
}
\seealso{
Other lookup_tables: 
\code{\link{country_thesaurus}},
\code{\link{db_info_table}}
}
\concept{lookup_tables}
