% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list.R
\name{order_variables}
\alias{order_variables}
\alias{order_variables.default}
\alias{order_variables.c14_date_list}
\title{Order variables (static key)}
\usage{
order_variables(x)

\method{order_variables}{default}(x)

\method{order_variables}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Order variables in a c14_date_list after a defined key
}
