% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_basic.R
\name{print.c14_date_list}
\alias{print.c14_date_list}
\title{Print a c14_date_list}
\usage{
\method{print}{c14_date_list}(x, ...)
}
\arguments{
\item{x}{a c14_date_list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print a c14_date_list according to \code{format.c14_date_list}.
The function uses the tibble print function and adds a custom
header for c14_date_lists.
}
\examples{
print(example_c14_date_list)

}
