% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_duplicates_remove.R
\name{remove_duplicates}
\alias{remove_duplicates}
\alias{remove_duplicates.default}
\alias{remove_duplicates.c14_date_list}
\title{remove duplicates}
\usage{
remove_duplicates(x)

\method{remove_duplicates}{default}(x)

\method{remove_duplicates}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list with the additional column duplicate_remove_log
}
\description{
remove double entries in a c14_date_list
}
