% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_rm_doubles.R
\name{rm_doubles}
\alias{rm_doubles}
\alias{rm_doubles.default}
\alias{rm_doubles.c14_date_list}
\title{Removes doubles}
\usage{
rm_doubles(x)

\method{rm_doubles}{default}(x)

\method{rm_doubles}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Removes double entries in a c14_date_list by
comparing the Labcodes
}
