% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_wrap_spatial.R
\name{spatial_country_checker}
\alias{spatial_country_checker}
\alias{spatial_country_checker.default}
\alias{spatial_country_checker.c14_date_list}
\title{Check and assign country based on spatial information}
\usage{
spatial_country_checker(x)

\method{spatial_country_checker}{default}(x)

\method{spatial_country_checker}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Check and assign country based on spatial information by first determining country named based on spatial join, and second compare to already supplied country information
}
