% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_duplicates_remove.R
\name{stringify_data_frame}
\alias{stringify_data_frame}
\title{stringify_data_frame}
\usage{
stringify_data_frame(x)
}
\arguments{
\item{x}{a data.frame}
}
\value{
a vector of strings describing the data.frame
}
\description{
stringify_data_frame
}
