% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list.R
\name{thesaurify}
\alias{thesaurify}
\alias{thesaurify.default}
\alias{thesaurify.c14_date_list}
\title{Apply thesaurus}
\usage{
thesaurify(x)

\method{thesaurify}{default}(x)

\method{thesaurify}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Add columns country_cor & material_cor with simplified and unified terms
}
