% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{variable_reference}
\alias{variable_reference}
\title{Variable Reference}
\format{A tibble.
\itemize{
  \item \bold{c14bazAAR} name of variable in c14bazAAR
  \item \bold{type} data type of variable in R
  \item \bold{definition} meaning of variable
  \item \bold{priority} priority of variable
  \item \bold{order} position of variable in a c14_date_list
  \item \bold{...} variables in source databases
}}
\description{
The parameter reference list of c14bazAAR: Which variables in
a c14_date_list equal the ones in the source databases and what
do they mean. Also contains a full list of the variables in the
source databases.
}
