% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_write_c14.R
\name{write_c14}
\alias{write_c14}
\alias{write_c14.default}
\alias{write_c14.c14_date_list}
\title{write \strong{c14_date_list}s to files}
\usage{
write_c14(x, format = c("csv"), ...)

\method{write_c14}{default}(x, format = c("csv"), ...)

\method{write_c14}{c14_date_list}(x, format = c("csv"), ...)
}
\arguments{
\item{x}{an object of class c14_date_list}

\item{format}{the output format: 'csv' (default) or 'xlsx'.
'csv' calls \code{utils::write.csv()},
'xlsx' calls \code{writexl::write_xlsx()}}

\item{...}{passed to the actual writing functions}
}
\description{
write \strong{c14_date_list}s to files
}
\examples{
write_c14(
  example_c14_date_list,
  file = tempfile(),
  format = "csv"
)

}
