% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14bazAAR.R
\docType{package}
\name{c14bazAAR-package}
\alias{c14bazAAR}
\alias{c14bazAAR-package}
\title{c14bazAAR: Download and Prepare C14 Dates from Different Source Databases}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Query different C14 date databases and apply basic data cleaning, merging and calibration steps. Currently available databases: 14cpalaeolithic, 14sea, adrac, agrichange, aida, austarch, bda, calpal, caribbean, eubar, euroevol, irdd, jomon, katsianis, kiteeastafrica, medafricarbon, mesorad, neonet, neonetatl, nerd, p3k14c, pacea, palmisano, rado.nb, rxpand, sard, tegiszhol, xronos.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/c14bazAAR}
  \item \url{https://github.com/ropensci/c14bazAAR}
  \item Report bugs at \url{https://github.com/ropensci/c14bazAAR/issues}
}

}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Dirk Seidensticker \email{dirk.seidensticker@gmail.com} (\href{https://orcid.org/0000-0002-8155-7702}{ORCID})
  \item Daniel Knitter \email{knitter@geographie.uni-kiel.de} (\href{https://orcid.org/0000-0003-3014-4497}{ORCID})
  \item Martin Hinz \email{martin.hinz@iaw.unibe.ch} (\href{https://orcid.org/0000-0002-9904-6548}{ORCID})
  \item David Matzig \email{matzigdavid@gmail.com} (\href{https://orcid.org/0000-0001-7349-5401}{ORCID})
  \item Wolfgang Hamer \email{hamer@geographie.uni-kiel.de} (\href{https://orcid.org/0000-0002-5943-5020}{ORCID})
  \item Kay Schmuetz \email{kschmuetz@ufg.uni-kiel.de}
}

Other contributors:
\itemize{
  \item Thomas Huet \email{thomashuet7@gmail.com} (\href{https://orcid.org/0000-0002-1112-6122}{ORCID}) [contributor]
  \item Nils Mueller-Scheessel \email{nils.mueller-scheessel@ufg.uni-kiel.de} (\href{https://orcid.org/0000-0001-7992-8722}{ORCID}) [contributor]
  \item Joe Roe \email{joe@joeroe.io} (\href{https://orcid.org/0000-0002-1011-1244}{ORCID}) [contributor]
  \item Ben Marwick (\href{https://orcid.org/0000-0001-7879-4531}{ORCID}) [reviewer]
  \item Enrico R. Crema (\href{https://orcid.org/0000-0001-6727-5138}{ORCID}) [reviewer]
}

}
\keyword{internal}
