% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14bazAAR.R
\docType{package}
\name{c14bazAAR-package}
\alias{c14bazAAR}
\alias{c14bazAAR-package}
\title{c14bazAAR: Download and Prepare C14 Dates from Different Source Databases}
\description{
Query different C14 date databases and apply basic data cleaning, merging and calibration steps.
}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} [copyright holder]

Authors:
\itemize{
  \item Wolfgang Hamer \email{hamer@geographie.uni-kiel.de}
  \item Martin Hinz \email{martin.hinz@ufg.uni-kiel.de}
  \item Daniel Knitter \email{knitter@geographie.uni-kiel.de}
  \item David Matzig
  \item Kay Schmütz \email{kschmuetz@ufg.uni-kiel.de}
  \item Dirk Seidensticker \email{dirk.seidensticker@gmail.com}
}

}
\keyword{internal}
