% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14bazAAR.R
\docType{package}
\name{c14bazAAR-package}
\alias{c14bazAAR}
\alias{c14bazAAR-package}
\title{c14bazAAR: Download and Prepare C14 Dates from Different Source Databases}
\description{
Query different C14 date databases and apply basic data cleaning, merging and calibration steps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ISAAKiel/c14bazAAR}
  \item Report bugs at \url{https://github.com/ISAAKiel/c14bazAAR/issues}
}

}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} [copyright holder]

Authors:
\itemize{
  \item Dirk Seidensticker \email{dirk.seidensticker@gmail.com}
  \item Daniel Knitter \email{knitter@geographie.uni-kiel.de}
  \item Martin Hinz \email{martin.hinz@ufg.uni-kiel.de}
  \item David Matzig
  \item Wolfgang Hamer \email{hamer@geographie.uni-kiel.de}
  \item Kay Schmütz \email{kschmuetz@ufg.uni-kiel.de}
}

Other contributors:
\itemize{
  \item Nils Mueller-Scheessel \email{nils.mueller-scheessel@ufg.uni-kiel.de} [contributor]
}

}
\keyword{internal}
