% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_calibrate.R
\name{calibrate}
\alias{calibrate}
\alias{calibrate.default}
\alias{calibrate.c14_date_list}
\title{Calibrate dates}
\usage{
calibrate(x)

\method{calibrate}{default}(x)

\method{calibrate}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Calibrate all valid dates in a c14_date_list
}
