% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_material_classification.R
\name{classify_material}
\alias{classify_material}
\alias{classify_material.default}
\alias{classify_material.c14_date_list}
\title{Apply material classification}
\usage{
classify_material(x, material_thesaurus_df = c14bazAAR::material_thesaurus)

\method{classify_material}{default}(x,
  material_thesaurus_df = c14bazAAR::material_thesaurus)

\method{classify_material}{c14_date_list}(x,
  material_thesaurus_df = c14bazAAR::material_thesaurus)
}
\arguments{
\item{x}{an object of class c14_date_list}

\item{material_thesaurus_df}{a thesaurus table (default: c14bazAAR::material_thesaurus)}
}
\value{
an object of class c14_date_list
}
\description{
Add column material_thes with simplified and unified terms
}
