% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_spatial_coordinate_precision.R
\name{coordinate_precision}
\alias{coordinate_precision}
\alias{coordinate_precision.default}
\alias{coordinate_precision.c14_date_list}
\title{Return coordinate precision according to number of digits in the columns
\strong{lat} and \strong{lon} of a \strong{c14_date_list}}
\usage{
coordinate_precision(x)

\method{coordinate_precision}{default}(x)

\method{coordinate_precision}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list with the additional column
\strong{coord_precision}
}
\description{
The precision of the coordinates for each date vary greatly.
\code{c14bazAAR::coordinate_precision()} calculates the mean of the possible
deviation in meters and adds it to the \strong{c14_date_list} with the column
\strong{coord_precision}.
}
\examples{
# calculate coordinate precision for all dates
ex <- coordinate_precision(example_c14_date_list)
ex[,c("lat", "lon", "coord_precision")]

}
