% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{country_thesaurus}
\alias{country_thesaurus}
\title{Country thesaurus}
\format{
...
}
\description{
Lookup table for country names.
}
\seealso{
Other lookup_tables: 
\code{\link{db_info_table}},
\code{\link{material_thesaurus}}
}
\concept{lookup_tables}
