% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_14sea.R, R/get_adrac.R, R/get_austarch.R,
%   R/get_c14data.R, R/get_calpal.R, R/get_context.R, R/get_eastafrica.R,
%   R/get_eubar.R, R/get_euroevol.R, R/get_irdd.R, R/get_palmisano.R,
%   R/get_radon.R, R/get_radonb.R
\name{get_dates}
\alias{get_dates}
\alias{get_14SEA}
\alias{get_aDRAC}
\alias{get_AustArch}
\alias{get_all_dates}
\alias{get_CalPal}
\alias{get_CONTEXT}
\alias{get_KITEeastAfrica}
\alias{get_EUBAR}
\alias{get_EUROEVOL}
\alias{get_IRDD}
\alias{get_Palmisano}
\alias{get_RADON}
\alias{get_RADONB}
\title{Backend functions for data download}
\usage{
get_14SEA(db_url = get_db_url("14SEA"))

get_aDRAC(db_url = get_db_url("aDRAC"))

get_AustArch(db_url = get_db_url("AustArch"))

get_all_dates()

get_CalPal(db_url = get_db_url("CalPal"))

get_CONTEXT(db_url = get_db_url("CONTEXT"))

get_KITEeastAfrica(db_url = get_db_url("KITEeastAfrica"))

get_EUBAR(db_url = get_db_url("EUBAR"))

get_EUROEVOL(db_url = get_db_url("EUROEVOL"))

get_IRDD(db_url = get_db_url("IRDD"))

get_Palmisano(db_url = get_db_url("Palmisano"))

get_RADON(db_url = get_db_url("RADON"))

get_RADONB(db_url = get_db_url("RADON-B"))
}
\arguments{
\item{db_url}{Character. URL that points to the c14 archive file. \code{c14bazAAR::get_db_url()}
fetches the URL from a reference list on github}
}
\description{
Backend functions to download data. See \code{?\link{get_c14data}}
for a more simple interface and further information.
}
