% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_14cpalaeolithic.R, R/get_14sea.R,
%   R/get_adrac.R, R/get_agrichange.R, R/get_austarch.R, R/get_bda.R,
%   R/get_c14data.R, R/get_calpal.R, R/get_caribbean.R, R/get_context.R,
%   R/get_eubar.R, R/get_euroevol.R, R/get_irdd.R, R/get_jomon.R,
%   R/get_katsianis.R, R/get_kiteeastafrica.R, R/get_medafricarbon.R,
%   R/get_mesorad.R, R/get_nerd.R, R/get_pacea.R, R/get_palmisano.R,
%   R/get_radon.R, R/get_radonb.R, R/get_rxpand.R
\name{get_14cpalaeolithic}
\alias{get_14cpalaeolithic}
\alias{get_14sea}
\alias{get_adrac}
\alias{get_agrichange}
\alias{get_austarch}
\alias{get_bda}
\alias{get_all_dates}
\alias{get_calpal}
\alias{get_caribbean}
\alias{get_context}
\alias{get_eubar}
\alias{get_euroevol}
\alias{get_irdd}
\alias{get_jomon}
\alias{get_katsianis}
\alias{get_kiteeastafrica}
\alias{get_medafricarbon}
\alias{get_mesorad}
\alias{get_nerd}
\alias{get_pacea}
\alias{get_palmisano}
\alias{get_radon}
\alias{get_radonb}
\alias{get_rxpand}
\title{Backend functions for data download}
\usage{
get_14cpalaeolithic(db_url = get_db_url("14cpalaeolithic"))

get_14sea(db_url = get_db_url("14sea"))

get_adrac(db_url = get_db_url("adrac"))

get_agrichange(db_url = get_db_url("agrichange"))

get_austarch(db_url = get_db_url("austarch"))

get_bda(db_url = get_db_url("bda"))

get_all_dates()

get_calpal(db_url = get_db_url("calpal"))

get_caribbean(db_url = get_db_url("caribbean"))

get_context(db_url = get_db_url("context"))

get_eubar(db_url = get_db_url("eubar"))

get_euroevol(db_url = get_db_url("euroevol"))

get_irdd(db_url = get_db_url("irdd"))

get_jomon(db_url = get_db_url("jomon"))

get_katsianis(db_url = get_db_url("katsianis"))

get_kiteeastafrica(db_url = get_db_url("kiteeastafrica"))

get_medafricarbon(db_url = get_db_url("medafricarbon"))

get_mesorad(db_url = get_db_url("mesorad"))

get_nerd(db_url = get_db_url("nerd"))

get_pacea(db_url = get_db_url("pacea"))

get_palmisano(db_url = get_db_url("palmisano"))

get_radon(db_url = get_db_url("radon"))

get_radonb(db_url = get_db_url("radonb"))

get_rxpand(db_url = get_db_url("rxpand"))
}
\arguments{
\item{db_url}{Character. URL that points to the c14 archive file. \code{c14bazAAR::get_db_url()}
fetches the URL from a reference list on github}
}
\description{
Backend functions to download data. See \code{?\link{get_c14data}}
for a more simple interface and further information.
}
