% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{deprecated_functions}
\alias{deprecated_functions}
\alias{mark_duplicates}
\alias{coordinate_precision}
\alias{finalize_country_name}
\alias{standardize_country_name}
\alias{get_emedyd}
\alias{fix_database_country_name}
\alias{classify_material}
\title{Deprecated functions}
\usage{
mark_duplicates(...)

coordinate_precision(...)

finalize_country_name(...)

standardize_country_name(...)

get_emedyd(...)

fix_database_country_name(...)

classify_material(...)
}
\arguments{
\item{...}{...}
}
\description{
Run them anyway to get some information about their replacements
or why they were removed.
}
