% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated_functions.R
\name{deprecated_functions}
\alias{deprecated_functions}
\alias{mark_duplicates}
\alias{coordinate_precision}
\alias{finalize_country_name}
\title{Deprecated functions}
\usage{
mark_duplicates(...)

coordinate_precision(...)

finalize_country_name(...)
}
\arguments{
\item{...}{...}
}
\description{
Run them anyway to get some information about their replacements
or why they were removed.
}
