% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_country_by_coordinate.R
\name{determine_country_by_coordinate}
\alias{determine_country_by_coordinate}
\alias{determine_country_by_coordinate.default}
\alias{determine_country_by_coordinate.c14_date_list}
\title{Determines country by coordinate}
\usage{
determine_country_by_coordinate(x)

\method{determine_country_by_coordinate}{default}(x)

\method{determine_country_by_coordinate}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Determines country by coordinate
}
