% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_spatial_coordinate_precision.R
\name{digits_counter}
\alias{digits_counter}
\title{digits_counter}
\usage{
digits_counter(x)
}
\arguments{
\item{x}{vector of coordinate values#'}
}
\value{
vecor with numer of digits
}
\description{
counts the digits of the given coordinates
}
