% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{example_c14_date_list}
\alias{example_c14_date_list}
\title{Example c14_date_list}
\format{
A c14_date_list. See data_raw/variable_definition.csv for an explanation of
the variable meaning.
}
\description{
c14_date_list with 1000 random dates for tests and example code.
}
\concept{c14_date_lists}
