% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_standardize_country_name.R
\name{find_lookup_decisions}
\alias{find_lookup_decisions}
\title{find_lookup_decisions}
\usage{
find_lookup_decisions(x, country_thesaurus)
}
\arguments{
\item{x}{a c14_date_list with country and country_thes}

\item{country_thesaurus}{data.frame with correct and variants of country names}
}
\value{
a tibble with the country names and the new country_thes names
found by \code{find_correct_name_by_stringdist_comparison()}
}
\description{
find_lookup_decisions
}
