% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_fuse.R
\name{fuse}
\alias{fuse}
\alias{fuse.default}
\alias{fuse.c14_date_list}
\title{Fuse multiple \strong{c14_date_list}s}
\usage{
fuse(...)

\method{fuse}{default}(...)

\method{fuse}{c14_date_list}(...)
}
\arguments{
\item{...}{objects of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
This function combines \strong{c14_date_list}s with
\code{dplyr::bind_rows()}. \cr
This is not a joining operation and it therefore
might introduce duplicates. See \code{c14bazAAR::mark_duplicates()}
and \code{c14bazAAR::remove_duplicates()} for a way to find and remove
them.
}
\examples{
# fuse three identical example c14_date_lists
fuse(example_c14_date_list, example_c14_date_list, example_c14_date_list)

}
