% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/14sea.R
\name{get_14SEA}
\alias{get_14SEA}
\title{get current 14SEA-Database}
\usage{
get_14SEA(db_url = get_db_url("14SEA"))
}
\arguments{
\item{db_url}{string with weblink to c14 archive file}
}
\description{
Downloads the current version of the 14SEA-Database from \url{http:///www.14sea.org/}.
}
\examples{

\dontrun{
  14SEA <- get_14SEA()
}

}
