% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adrac.R
\name{get_aDRAC}
\alias{get_aDRAC}
\title{get current aDRAC-Database}
\usage{
get_aDRAC()
}
\description{
Downloads the current version of the aDRAC-Database from \url{https://github.com/dirkseidensticker/aDRAC/}.
}
\examples{

\dontrun{
  aDRAC <- get_aDRAC()
}

}
