% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/austarch.R
\name{get_AustArch}
\alias{get_AustArch}
\title{get AustArch: A Database of 14C and Luminescence Ages from Archaeological Sites in Australia}
\usage{
get_AustArch(db_url = get_db_url("AustArch"))
}
\arguments{
\item{db_url}{string with weblink to c14 archive file}
}
\description{
See:
- AustArch: A Database of 14C and Non-14C Ages from Archaeological Sites in Australia - Composition, Compilation and Review (Data Paper) http://intarch.ac.uk/journal/issue36/6/williams.html
- AustArch: A Database of 14C and Luminescence Ages from Archaeological Sites in Australia http://archaeologydataservice.ac.uk/archives/view/austarch_na_2014/
}
\examples{

\dontrun{
  AustArch <- get_AustArch()
}

}
