% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calpal.R
\name{get_CalPal}
\alias{get_CalPal}
\title{get current CALPAL-Database}
\usage{
get_CalPal(db_url = get_db_url("CalPal"))
}
\arguments{
\item{db_url}{string with weblink to c14 archive file}
}
\description{
Downloads an (actually old) version of the CALPAL-Database from \url{https://github.com/nevrome/CalPal-Database}.
}
\examples{

\dontrun{
  CALPAL <- get_CALPAL()
}

}
