% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context.R
\name{get_CONTEXT}
\alias{get_CONTEXT}
\title{get current CONTEXT-Database}
\usage{
get_CONTEXT()
}
\description{
Downloads the current version of the CONTEXT-Database from \url{http://context-database.uni-koeln.de/}.
}
\examples{

\dontrun{
  CONTEXT <- get_CONTEXT()
}

}
