% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_db_urls.R
\name{get_db_url}
\alias{get_db_url}
\title{get db url}
\usage{
get_db_url(
  db_name,
 
    ref_url = "https://raw.githubusercontent.com/ropensci/c14bazAAR/master/data-raw/url_reference.csv"
)
}
\arguments{
\item{db_name}{name of the database}
}
\description{
Downloads URLs of c14 source databases from a reference table
on github.
}
