% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euroevol.R
\name{get_EUROEVOL}
\alias{get_EUROEVOL}
\title{get current EUROEVOL-Database}
\usage{
get_EUROEVOL()
}
\description{
Downloads the current version of the EUROEVOL-Database from \url{http://discovery.ucl.ac.uk/1469811/}.
}
\examples{

\dontrun{
  EUROEVOL <- get_EUROEVOL()
}

}
