% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radon.R
\name{get_RADON}
\alias{get_RADON}
\title{get current RADON-Database}
\usage{
get_RADON(db_url = get_db_url("RADON"))
}
\arguments{
\item{db_url}{string with weblink to c14 archive file}
}
\description{
Downloads the current version of the RADON-Database from \url{http://radon.ufg.uni-kiel.de/}.
}
\examples{

\dontrun{
  RADON <- get_RADON()
}

}
