% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_spatial_coordinate_precision.R
\name{individual_precision}
\alias{individual_precision}
\title{individual_precision}
\usage{
individual_precision(x, mode)
}
\arguments{
\item{x}{vector of coordinates (latitude or longitude)}

\item{mode}{argument indicating the `mode` of the coordinates (whether these are lat or lon)}
}
\value{
vector with precision in meters
}
\description{
individual_precision
}
