% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinate_precision.R
\name{individual_precision}
\alias{individual_precision}
\title{individual_precision}
\usage{
individual_precision(x, mode)
}
\arguments{
\item{x}{vector of coordinates (latitude or longitude)}

\item{...}{further arguments indicating the `mode` of the coordinates (whether these are lat or lon)}
}
\value{
vecor with precision in meters
}
\description{
individual_precision
}
