% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_thesauri.R
\name{inspect_lookup_country}
\alias{inspect_lookup_country}
\alias{inspect_lookup_material}
\title{Get an overview of the lookup decisions}
\usage{
inspect_lookup_country(x, thesaurus = c14bazAAR::country_thesaurus)

inspect_lookup_material(x, thesaurus = c14bazAAR::material_thesaurus)
}
\arguments{
\item{x}{a c14_date_list with the columns country and country_thes or material
and material_thes which went through either or both \link{fix_database_country_name}
or \link{classify_material}}

\item{thesaurus}{see \link{fix_database_country_name} or \link{classify_material}}
}
\value{
a data.frame with information about lookup decisions. Columns:
\itemize{
 \item{already_in_thesaurus: entry is already identical to the "correct" one in the thesaurus table}
 \item{changed: entry was changed with \link{fix_database_country_name} or \link{classify_material}}
 \item{changed_thesaurus: entry was changed with a lookup in the thesaurus}
 \item{changed_not_thesaurus: entry was changed by lookup in another reference}
 \item{not_changed_not_thesaurus: entry was not changed is not in the thesaurus}
}
For material classification this last column indicates
values that should be added to the thesaurus, because
they are missing there. For countries it often means
that the entries were found in \link[countrycode]{codelist}.
}
\description{
These functions allow to inspect the lookups done in \link{fix_database_country_name}
and \link{classify_material}. This can be useful to understand which values are not
covered in the in the thesaurus tables.
}
