% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_standardize_country_name.R
\name{lookup_in_countrycode_codelist}
\alias{lookup_in_countrycode_codelist}
\title{lookup_in_countrycode_codelist}
\usage{
lookup_in_countrycode_codelist(x, codesets, manual_attribution, ...)
}
\arguments{
\item{x}{vector of country codes to look up in countrycode codelist}

\item{codesets}{which country codesets should be searched beyond "country.name.en"}

\item{manual_attribution}{named vector. correct country names and the wrong values in country column}

\item{...}{additional arguments are passed to stringdist::stringdist()}
}
\value{
a vector with the correct english country names
}
\description{
lookup_in_countrycode_codelist
}
