% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_duplicates_mark.R
\name{mark_duplicates}
\alias{mark_duplicates}
\alias{mark_duplicates.default}
\alias{mark_duplicates.c14_date_list}
\title{mark duplicates}
\usage{
mark_duplicates(x)

\method{mark_duplicates}{default}(x)

\method{mark_duplicates}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list with the additional column duplicate_group
}
\description{
mark double entries in a c14_date_list by
comparing the Labcodes
}
