% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{material_thesaurus}
\alias{material_thesaurus}
\title{Material Thesaurus}
\format{A tibble with 2 variables.
\itemize{
  \item \bold{cor} fixed name
  \item \bold{var} variations
}}
\description{
A small thesaurus for material classes.
}
\seealso{
Other thesauri: \code{\link{country_thesaurus}}
}
\concept{thesauri}
