% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_order_variables.R
\name{order_variables}
\alias{order_variables}
\alias{order_variables.default}
\alias{order_variables.c14_date_list}
\title{Order the variables in a \strong{c14_date_list}}
\usage{
order_variables(x)

\method{order_variables}{default}(x)

\method{order_variables}{c14_date_list}(x)
}
\arguments{
\item{x}{an object of class c14_date_list}
}
\value{
an object of class c14_date_list
}
\description{
Order variables following a key defined in the column
\strong{order} in \code{c14bazAAR::variable_reference}. This makes
sure that a \strong{c14_date_list} always appears with the same
outline. \cr
A \strong{c14_date_list} has at least the columns \strong{c14age}
and \strong{c14std}. Beyond that there's a selection of additional
variables depending on the input from the source databases, as a
result of the \code{c14bazAAR} functions or added by other data
analysis steps. This function arranges the expected variables in
a distinct, predefined order. Undefined variables are added at the
end.
}
