% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list.R
\name{print.c14_date_list}
\alias{print.c14_date_list}
\title{Print a c14_date_list}
\usage{
\method{print}{c14_date_list}(x, ...)
}
\arguments{
\item{x}{a c14_date_list}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print a c14_date_list according to the format.c14_date_list
}
