% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_standardize_country_name.R
\name{print_lookup_decisions}
\alias{print_lookup_decisions}
\title{print_lookup_decisions}
\usage{
print_lookup_decisions(x, country_thesaurus)
}
\arguments{
\item{x}{a c14_date_list with country and country_thes}

\item{country_thesaurus}{data.frame with correct and variants of country names}
}
\value{
NULL, called for the print side effect
}
\description{
print_lookup_decisions
}
