% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_rm_doubles.R
\name{rm_doubles}
\alias{rm_doubles}
\alias{rm_doubles.default}
\alias{rm_doubles.c14_date_list}
\title{Removes doubles}
\usage{
rm_doubles(x, mark = FALSE)

\method{rm_doubles}{default}(x, mark = FALSE)

\method{rm_doubles}{c14_date_list}(x, mark = FALSE)
}
\arguments{
\item{x}{an object of class c14_date_list}

\item{mark}{should the low quality elements of a double entry
collection only be marked in a new column \code{low_qual} instead
of being removed? Default: FALSE}
}
\value{
an object of class c14_date_list
}
\description{
Removes double entries in a c14_date_list by
comparing the Labcodes
}
