% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c14_date_list_thesaurify.R
\name{thesaurify}
\alias{thesaurify}
\alias{thesaurify.default}
\alias{thesaurify.c14_date_list}
\title{Apply thesaurus}
\usage{
thesaurify(x, country_thesaurus_table = c14databases::country_thesaurus,
  material_thesaurus_table = c14databases::material_thesaurus)

\method{thesaurify}{default}(x,
  country_thesaurus_table = c14databases::country_thesaurus,
  material_thesaurus_table = c14databases::material_thesaurus)

\method{thesaurify}{c14_date_list}(x,
  country_thesaurus_table = c14databases::country_thesaurus,
  material_thesaurus_table = c14databases::material_thesaurus)
}
\arguments{
\item{x}{an object of class c14_date_list}

\item{country_thesaurus_table}{a thesaurus table (default: c14databases::country_thesaurus)}

\item{material_thesaurus_table}{a thesaurus table (default: c14databases::material_thesaurus)}
}
\value{
an object of class c14_date_list
}
\description{
Add columns country_cor & material_cor with simplified and unified terms
}
